Move SVD into its own module
